ml/requirements.txt
 # EmotionEcho ML Requirements
# Python 3.10+

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers & Models
transformers>=4.30.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=9.5.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress & Logging
tqdm>=4.65.0

# Experiment Tracking (Optional)
mlflow>=2.8.0

# Data Validation (Optional)
pydantic>=2.0.0

# Testing (Optional)
pytest>=7.4.0
pytest-cov>=4.1.0